fix xoa hop dong don le theo mau

+++ src/app/main/contract/form-contract/party-contract-form/party-contract-form.component.ts
@@ -1585,8 +1585,9 @@ export class PartyContractFormComponent implements OnInit, AfterViewInit {
 
   // xóa đối tham gia bên đối tác
   deletePartner(index: any, item: any) {
+    console.log("item ", item);
     //xoa doi tuong tham gia
-    if (item.id) {
+    if (item.id && this.router.url.includes('edit')) {
       this.spinner.show();
       this.contractService.deleteParticipantContract(item.id).subscribe((res: any) => {
         if (res.success == true) {
